ENH: Allow sliding session expiration
In most webapps it is an anti-pattern/security risk to have sliding
session expirations based upon requests. In our use case, however, with
shared same-site resources, it makes sense to. This ensures new sessions
do not need to be recreated as long as the user is active and allows the
GUI to think less about session expiration itself.