FIX: Store root proxy with project as RMS session
The lifetime of the worker is controlled through the root proxy only.
This means a reference to it should be stored alongside the project so
that shutdown can be called on it. Some adjustments to the usage of the
RMS service and session service came along to make this more clearly
handled with the services.